src/main/java/zhangyi/refactoring/mfbook/model/Customer.java
package zhangyi.refactoring.mfbook.model;

import java.util.ArrayList;
import java.util.List;

public class Customer {
    private String name;
    private List<Rental> rentals = new ArrayList<>();

    public Customer(String name) {
        this.name = name;
    }

    public void addRental(Rental rental) {
        rentals.add(rental);
    }

    public String getName() {
        return name;
    }

    public int getTotalPoints() {
        int frequentRenterPoints = 0;
        for (Rental rental : rentals) {
            frequentRenterPoints += rental.getFrequentRenterPoints();
        }
        return frequentRenterPoints;
    }

    public double getTotalAmount() {
        double totalAmount = 0;
        for (Rental rental : rentals) {
            double thisAmount = rental.getCharge();
            totalAmount += thisAmount;
        }
        return totalAmount;
    }

    public List<Rental> getRentals() {
        return rentals;
    }
}
